+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 4.1.63 - 2026-01-24T17:44:59.598Z
+Previous: 4.1.62
+Commit: 13ecd18
+Changes:
+- No commits detected since last build
+
+## 4.1.62 - 2026-01-24T17:43:56.248Z
+Previous: 4.1.61
+Commit: 13ecd18
+Changes:
+- Widen message column to 50
+
 ## 4.1.61 - 2026-01-24T17:35:45.825Z
 Previous: 4.1.60
 Commit: 065feda